Changes to Summary File
Updated the code to automatically sort the each file by the filename. Also added some code to normalize the relative amplitude in each channel the to lowest relative amplitude number.

Noticed an issue with multichannel movies where the relative amplitude number is extremely high for some movies. This is becuase the min number of some of the boxes is very low. I'm not sure if this is an issue or not