Update
Fixed a bug where if tried to do group comparisons with one channel movies, the code would try to add shift stats.

Also added a function called: ave_means_to_csv. It saves the means of the main wave categories so that it is easier to copy and paste to prism