Saving CCF values for future plotting
In this version, I included the feature to store the average CCF values for each channel combination. I also incorporated the optional feature to save the individual CCF values for each box (for all combinations). This is especially beneficial when aiming to create visually appealing graphics for publishing or presentations.

Furthermore, I removed the sample data as it is unnecessary for my specific needs.