"fix" small delay issue
if the shift between signals is very small, sometimes the script will choose the second furthest shift form zero, which is just the period of the wave. Therefore, I added in a bit of code that will detect if the shift is larger than 60% of the period, and will subtract one whole period from that shift value.